finishing home screen

+++ app/src/main/java/com/example/climatic/model/responses/WeatherResponse.kt
@@ -19,7 +19,7 @@ data class WeatherResponse(
     @SerializedName("visibility") var visibility: Int? = null,
     @SerializedName("wind") var wind: Wind? = Wind(),
     @SerializedName("clouds") var clouds: Clouds? = Clouds(),
-    @SerializedName("dt") var dt: Int? = null,
+    @SerializedName("dt") var dt: Long? = null,
     @SerializedName("sys") var sys: Sys? = Sys(),
     @SerializedName("timezone") var timezone: Int? = null,
     @SerializedName("id") var id: Int? = null,
